requirements/streamlit.txt
streamlit
plotly
snowflake-connector-python[pandas]
python-dotenv
